float instead of Int32 for regulation
With FPU, PID execution time is 0.5uS bigger (7.5->8uS)=> no problem at
all

+++ Sources/tasks/gFlightCompute.h
@@ -23,9 +23,9 @@ typedef enum {
 
 //Struct that contains the PID datas for each PID
 typedef struct {
-	Int32 aKp, aKi, aKd;
-	Int32 aPreviousError;
-	Int32 aIntegral;
+	float aKp, aKi, aKd;
+	float aPreviousError;
+	float aIntegral;
 	UInt16 aPreviousTime;
 } PIDdata;
 PIDdata sPID [3];
